// packages/react/src/components/shared/breadcrumb/common-breadcrumb.types.ts
import type { ReactElement, ReactNode } from "react";

export type BreadcrumbItemsConfig =
  | {
      id: string | number;
      label: string;
      href: string;
      render?: never;
      icon?: ReactNode;
    }
  | {
      id: string | number;
      label: string;
      href?: never;
      render: ReactElement;
      icon?: ReactNode;
    };

export type EllipsisItemConfig = {
  id: string;
  isEllipses: boolean;
  content: BreadcrumbItemsConfig[];
};

export interface CommonBreadCrumbsProps {
  items: BreadcrumbItemsConfig[];
  maxVisible?: number;
}